+++ srcs/Channel.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Channel.cpp                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ldesnoye <ldesnoye@student.42.fr>          +#+  +:+       +#+        */
+/*   By: aessakhi <aessakhi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/16 17:26:20 by ldesnoye          #+#    #+#             */
-/*   Updated: 2023/02/21 11:45:06 by ldesnoye         ###   ########.fr       */
+/*   Updated: 2023/02/21 13:27:18 by aessakhi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ bool	Channel::protectedTopicMode() const
 bool	Channel::noExternalMessagesMode() const
 { return _no_external_messages_mode ; }
 
-int		Channel::capacity() const
+size_t	Channel::capacity() const
 { return _capacity ; }
 
 /* -----ADDING USERS TO LISTS----- */
